requirements.txt
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
dj-database-url==3.0.1
django==5.2.4
django-cors-headers==4.7.0
django-cors-middleware==1.5.0
djangorestframework==3.16.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.4
whitenoise==6.9.0